:truck: domain package 이동 및타입 수정
