+++ src/main/kotlin/lotto/model/LottoNumbers.kt
@@ -1,22 +1,5 @@
 package lotto.model
 
-import lotto.domain.LottoPrizeResults
-
-data class Lotto(
-    val numbers: List<LottoNumbers>,
-) {
-    val size: Int
-        get() = numbers.size
-
-    fun matches(winningLotto: WinningLotto): LottoPrizeResults =
-        numbers.mapNotNull(winningLotto::prize)
-            .let(::LottoPrizeResults)
-
-    companion object {
-        const val PRICE = 1000
-    }
-}
-
 data class LottoNumbers(
     val value: List<LottoNumber>,
 ) {
