+++ src/test/kotlin/lotto/domain/LottoTest.kt
@@ -1,6 +1,8 @@
-package lotto.model
+package lotto.domain
 
-import org.assertj.core.api.Assertions
+import lotto.model.LottoNumber
+import lotto.model.LottoNumbers
+import lotto.model.LottoPrize
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
